Add a sysctl to show the pmap of a process
This adds a couple of sysctls that let you specify a process and
(optionally) a virtual address within that process, and another
sysctl that lets you display the corresponding physical page map.

There are a couple of CPU registers, MSRs, & CPUIDs that control/report
how memory management is done.    The debug.dump_pmap sysctl includes
these.

The sysctls don't use sbuf_printf() directly to render the string,
they take a passed-in function pointer to do the printing.  This
is so they can be used in ddb too (see the following commit).