Fixes: LRS-3350 (Add a sysctl to show the vmspace of a process)
Problem: When adding 1 GB superpages, it'd be useful to be able to
inspect the physical page map of a process.  This ticket is to add
that insight, via a sysctl.

Analysis:  This adds a couple of sysctls that let you specify a
process and (optionally) a virtual address within that process, and
another sysctl that lets you display the corresponding physical
page map.

Unit/Robot tests:  Ran by hand in the test lab, normal and Full
Debug builds.