Add a sysctl to show the pmap of a process
This adds a couple of sysctls that let you specify a process and
(optionally) a virtual address within that process, and another
sysctl that lets you display the corresponding physical page map.

There are a couple of CPU registers, MSRs, & CPUIDs that control/report
how memory management is done.    The debug.dump_pmap sysctl includes
these.