Implement kmem_alloc_contig() in the 1 gig buddy allocator
Some in-kernel allocations go through the kmem_alloc_contig()
function (in addition to the allocations that go through kmem_malloc()),
to get physically contiguous memory.

The 1 gig allocator does not implement the kmem_alloc_contig()
function, so those allocations were satisfied by the old allocator,
which uses smaller pages.

Memory on a 1 gig page is already physically contiguous, this is
required by the pmap.  So satisfying allocation requests for
physically contiguous memory is easy.

A wrinkle is that kmem_alloc_contig() lets the caller specify upper
and lower limits on the acceptable physical address.  In practice
this is not a problem, because all current callers that i spot-checked
set the lower limit to 0 and the upper limit to the highest address.