teach dump_pmap sysctl to show kernel memory
It would sometimes be handy to inspect the pmap of kernel addresses,
but there's currently no easy way to do that.

There is a sysctl that shows the pmap of a pointer in a process,
but it doesn't know how to deal with the kmem_pmap.

Augment that sysctl so that if the pid is <= 0, it shows pointers
in the kmem_pmap instead.