make demotion of 1gig pages work correctly
pmap_protect() currently calls panic() when asked to change
protection of memory that lives on a 1 gig superpage.  This
does not normally happen, but there are situations where it
might, and we should not panic in that case.

Fix up pmap_protect() to correctly demote 1 gig superpages,
instead of panicking.