show the requested memory in dump_pmap sysctl
The sysctl that shows the pmap of a pointer (in the kernel or in a
userspace process) includes a hexdump of the leaf-node of the page
map, but it shows the first few bytes of the page instead of the
bytes pointed to by the pointer.

Change the sysctl to show the requested pointer.