Optimize setting of not_available bit
When a chunk is allocated, the bits for all of its children need
to be set as well. Currently, this requires setting each individual
bit in a loop. When a large chunk is allocated, this can be up to
1M bits

Adapt the freebsd bit_setn ans bit_clearn for 64bit words, and use
to set and clear continguous bits