Start searching for free blocks at cached word
The buddy allocator is designed to fill starting from the lowest
addresses. As chunks of some size are allocated, the next free chunk
is requires that more of the bitstring is searched

Save the word of the block that was last modified, so that all bits
for lower addresses are known to be not available.